+++ bmds/stats/anova.py
